fix calculating the last timestamp from query results
Now that one goroutine can fetch data from multiple streams in one pull,
returned results may not be sorted by their timestamps. When calculating
the largest timestamp of the returned results, it is not sufficient to
take the value of the last row, instead a real maximum value must be
chosen.