compare timestamps properly
Each worker goroutine keeps track of the biggest timestamp polled from
the cdc log so far, so that we won't read old rows in subsequent polls.
The timestamp was by comparing timeuuid lexicographically by their
bytes, which is naive and wrong. This caused the maximum not to be
updated for long periods of time, which caused cdc-stressor to read an
increasing and unnecessary amount of rows, which puts too much stress on
the cluster.

This patch fixes the comparison function and uses the same algorithm as
Cassandra and Scylla.